backend/requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
passlib[argon2]
python-jose
python-multipart
scikit-learn
pandas
numpy
pyportfolioopt
cohere
python-dotenv
joblib
yfinance
textblob
nltk
argon2-cffi
requests